switched to flipping data from flipping header
The approach of changing the header to say that the data is flipped is
flawed because how how cmtk handles space. It is too difficult to align
these flipped images to the template. Therefore, segment-neuropil makes
uses np.flip to actually flip the data. Although this has a significant
performance disadvantage, it makes the code much more simple